requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.15
django-cors-headers==3.2.1
django-redis==4.11.0
django-redis-cache==2.1.1
djangorestframework==3.11.0
gunicorn==20.0.2
idna==2.9
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
django-cors-middleware